+++ R/PersonClass.R
@@ -12,13 +12,15 @@ Person <- setRefClass("Person",
                                     item_time = 'numeric',
                                     valid_item = 'logical',
                                     login_name = 'character',
-                                    score = 'logical'),
+                                    score = 'logical',
+                                    true_thetas = 'numeric'),
                       
                       methods = list(
                          initialize = function(nfact, nitems, thetas.start_in, score,
                                                theta_SEs, ID = 0L){
                              'Initialize the person object given background information'
                              ID <<- ID
+                             true_thetas <<- numeric(0L)
                              raw_responses <<- as.character(rep(NA, nitems))
                              responses <<- as.integer(rep(NA, nitems))
                              valid_item <<- rep(TRUE, nitems)