+++ R/zzz-methods.R
@@ -6,9 +6,14 @@ print.mirtCAT <- function(x, ...){
     if(!all(is.na(x$thetas))){
         person <- c(sum(!is.na(x$raw_responses)),
                           x$thetas[1L,],
-                          x$SE_thetas[1L,])        
+                          x$SE_thetas[1L,])
         names(person) <- c('n.items.answered', paste0('Theta_', 1:length(x$thetas)),
                            paste0('SE.Theta_', 1:length(x$thetas)))
+        if(!is.null(x$true_thetas)){
+            nms <- names(person)
+            person <- c(person, x$true_thetas)
+            names(person) <- c(nms, paste0('True.Theta_', 1:length(x$thetas)))
+        } 
         ret <- t(as.data.frame(person))
         rownames(ret) <- ''
         print(ret)
@@ -43,6 +48,7 @@ summary.mirtCAT <- function(object, sort = TRUE, ...){
                 thetas_history=object$thetas_history, 
                 thetas_SE_history=object$thetas_SE_history,
                 demographics=object$demographics)
+    if(!is.null(object$true_thetas)) ret$true_thetas <- object$true_thetas
     if(!length(object$login_name)) ret$login_name <- NULL
     if(is.null(person)) ret$final_estimates <- NULL
     if(all(is.na(scored_responses))) ret$scored_responses <- NULL
@@ -63,7 +69,10 @@ summary.mirtCAT <- function(object, sort = TRUE, ...){
 #' @method plot mirtCAT
 #' @param pick_theta a number indicating which theta to plot (only applicable for multidimensional 
 #'   tests). The default is to facet each theta on one plot, but to plot only the first factor pass
-#'   \code{pick_theta = 1}   
+#'   \code{pick_theta = 1}
+#' @param true_thetas logical; include a horizontal line indicating where the population-level 
+#'   theta values are? Only applicable to Monte Carlo simulations because this value would not 
+#'   be known otherwise   
 #' @param main title of the plot. Will default to \code{'CAT Standard Errors'} or 
 #'   \code{'CAT ##\% Confidence Intervals'} depending on the SE input
 #' @param SE size of the standard errors to plot. The default is 1, and therefore plots the
@@ -71,7 +80,7 @@ summary.mirtCAT <- function(object, sort = TRUE, ...){
 #' @param par.strip.text plotting argument passed to \code{\link{lattice}}
 #' @param par.settings plotting argument passed to \code{\link{lattice}}
 #' @export
-plot.mirtCAT <- function(x, pick_theta = NULL, SE = 1, main = NULL, 
+plot.mirtCAT <- function(x, pick_theta = NULL, true_thetas = TRUE, SE = 1, main = NULL, 
                          par.strip.text = list(cex = 0.7),
                          par.settings = list(strip.background = list(col = '#9ECAE1'),
                                              strip.border = list(col = "black")),
@@ -115,7 +124,8 @@ plot.mirtCAT <- function(x, pick_theta = NULL, SE = 1, main = NULL,
                       main = main,
                       lower = thetasSElowlong$F1,
                       upper = thetasSEhighlong$F1,
-                      panel = function(x, y, lower, upper, subscripts, ...){
+                      panel = function(x, y, lower, upper, subscripts, true_thetas, 
+                                       true_theta_values, ...){
                           lower <- lower[subscripts]; upper <- upper[subscripts]
                           panel.polygon(c(x, rev(x)), c(upper, rev(lower)), 
                                         col="#E6E6E6", border = FALSE, ...)
@@ -125,28 +135,36 @@ plot.mirtCAT <- function(x, pick_theta = NULL, SE = 1, main = NULL,
                               tmp <- CV[subscripts]
                               panel.abline(h=tmp[1], col = 'red', ...)
                           }
+                          if(true_thetas && length(true_theta_values))
+                              panel.abline(h = true_theta_values[subscripts][1L], col = 'red', ...)
                       },
                       ylim=c(min(thetasSElowlong$F1)-.1, max(thetasSEhighlong$F1)+.1),
                       ylab = expression(theta), 
                       xlab = 'Item', par.strip.text=par.strip.text,
-                      par.settings=par.settings, ...))
+                      par.settings=par.settings, 
+                      true_theta_values = rep(x$true_thetas, each=nrow(thetaslong)/nfact), 
+                      true_thetas=true_thetas, ...))
         
     } else {
         return(xyplot(F1 ~ item, data=thetaslong, 
                       main = main,
                       lower = thetasSElowlong$F1,
                       upper = thetasSEhighlong$F1,
-                      panel = function(x, y, lower, upper, subscripts, ...){
+                      panel = function(x, y, lower, upper, subscripts, true_thetas, 
+                                       true_theta_values, ...){
                           panel.polygon(c(x, rev(x)), c(upper, rev(lower)), 
                                         col="#E6E6E6", border = FALSE, ...)
                           panel.points(x, y, ...)
                           panel.lines(x, y, ...)
                           if(!is.null(CV))
                               panel.abline(h=CV[1], col = 'red', ...)
+                          if(true_thetas && length(true_theta_values))
+                              panel.abline(h = true_theta_values, col = 'red', ...)
                       },
                       ylim=c(min(thetasSElowlong$F1)-.1, max(thetasSEhighlong$F1)+.1),
                       ylab = expression(theta), 
                       xlab = 'Item', par.strip.text=par.strip.text,
-                      par.settings=par.settings, ...))
+                      par.settings=par.settings, 
+                      true_theta_values = x$true_thetas, true_thetas=true_thetas, ...))
     }
 }